feat: split build tasks into files
