+++ scripts/constants.ts
@@ -1,0 +1,11 @@
+export const CONTINENTS = 'continents'
+export const COUNTRIES = 'countries'
+export const LANGUAGES = 'languages'
+
+export const ALL = '.all'
+
+export const CSV_EXT = 'csv'
+
+export const COMMA = '","'
+export const LF = '\n'
+export const QUOTE = '"'
